protobuf-c: eliminate void pointer arithmetic
gcc silently treats arithmetic on a void pointer as char pointer
arithmetic without -pedantic. Other compilers do not. Make it explicit
that we are doing arithmetic on a pointer with object size 1.

This eliminates the following diagnostic:

    protobuf-c/protobuf-c.c: In function 'parse_packed_repeated_member':
    protobuf-c/protobuf-c.c:2404:34: warning: pointer of type 'void *' used in arithmetic [-Wpedantic]
      void *array = *(void **) member + siz * (*p_n);

(I did not find any other instances of 'void *' arithmetic when
compiling with -pedantic.)

Based on a patch from alex85k (#167).