Add meson build system
This commit adds what should be a feature complete meson build system.

The only real difference with the existing autotools build system is
that, for the protobuf-c compiler, we now ship "protoc-c" as the actual
binary, and "protoc-gen-c" (the plugin for protoc) as a symlink to the
compiler, rather than the opposite. This is because meson doesn't have
native support for shipping symlinks, and I could not find a way to add
dependency edges from the generated .pb-c.[ch] files to the built
compiler when the built compiler binary was "protoc-gen-c" the compiler
was being invoked indirectly via something like "protoc
--plugin=protoc-gen-c=...".

There are two meson build options, "build-compiler" and "build-docs".
build-compiler controls whether the protoc-c binary is built (and like
the autotools build system, this is required to run the test suite), and
build-docs controls whether the Doxygen HTML documentation is built.