Set rpath on cxx-generate-packed-data so we don't need to set LD_LIBRARY_PATH
This commit sets the 'build_rpath' argument on the
cxx-generate-packed-data executable to the protobuf library's libdir,
which supports the use case where the protobuf library is installed into
a non-default path. For instance, in the CI environment protobuf is
installed into a subdirectory of $HOME, which is not searched by the
runtime linker.

This approach was recommended on the mesonbuild mailing list.