TL;DR: Added support for linux kernel.
Added makefile that compiles protobuf-c.ko (similiar to libprotobuf-c.a).
After protobuf-c.ko is loaded, kernel code that uses protobuf can run -
all that is needed is to include the header generated by protoc.
A major use case is replacing the NFS XDR encoding for example,
but any kernel module that wants to efficiently serialize data
to another machine (possibly also same machine if the data is large,
sparse...)

This PR does a fix which IMO is necessary regardless of the kernel support
which is to use system_free in protobuf_c_buffer_simple_clear (in
stead of just calling free(), so if you would like,
I will extract only this part to a separate PR.

I also have an example kernel module which uses protobuf in kernel,
but I don't know where to put it (some kind of examples dir?)