Fixed compiler warning on static init oneof field
The oneof field is a union and initialized by "0" in generated header file. But
the protobuf type "bytes" is stored using ProtobufCBinaryData. If bytes is the
first field of the "oneof", the "oneof" should be initialized by "{0,NULL}".

Some compilers report warning if "-Wall" is enabled:

In file included from test.c:1:
test.pb-c.h:43:2: warning: missing braces around initializer [-Wmissing-braces]
   43 |  { PROTOBUF_C_MESSAGE_INIT (&base__descriptor) \
      |  ^
test.c:3:10: note: in expansion of macro ‘BASE__INIT’
    3 | Base x = BASE__INIT;
      |          ^~~~~~~~~~

And if "-Werror" is enabled, the source code will not compile.

This patch fixed this issue by initializing the "oneof" with the first field.

Fixes: af613f932 ("Fix union initialization")
Signed-off-by: Zhai Zhaoxuan <kxuanobj@gmail.com>