Add generator parameter to disable message pack/unpack wrappers
This prevents the generation of specific helper functions for
get_packed_size, pack, pack_to_buffer, unpack and free_unpacked for each
message type. The generic protobuf_c_message version of these can still
be called. This significantly reduces the size of the generated libraries.

The command to do this looks like:
 protoc --c_out=disable_message_helpers:outdir foo.proto

Signed-off-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>