Makefile.am: Remove conditional BUILD_PROTO3 rules
Now that we require protobuf >= 3.0.0, we are always building with
proto3 syntax support.

Note that the test-generated-code3 test keeps `-DPROTO3` which is
separate from the HAVE_PROTO3 identifier. This test is building built
from the same source file as test-generated-code but the source file
supports proto2 and proto3 syntax depending on how it's compiled.