protobuf-c: Zero memory returned by do_alloc()
There have been reports of memory corruption, for instance the GitHub
issues #690 and #745, that are likely due to failure to fully initialize
allocated memory along a particular code path.

I have not been able to track down the immediate cause of #745
which is invoked by new optimizations in gcc 15 that are enabled by
`-fzero-init-padding-bits=standard`.

Clearing all allocated memory should be safe and any slight performance
hit is certainly worth it. It does appear to prevent the crash in #745.